moved spark folder into backends
