+++ ibis/backends/spark/api.py
@@ -1,7 +1,7 @@
 import ibis.common.exceptions as com
-from ibis.spark.client import SparkClient
-from ibis.spark.compiler import dialect  # noqa: F401
-from ibis.spark.udf import udf  # noqa: F401
+from ibis.backends.spark.client import SparkClient
+from ibis.backends.spark.compiler import dialect  # noqa: F401
+from ibis.backends.spark.udf import udf  # noqa: F401
 
 
 def compile(expr, params=None):
@@ -12,7 +12,7 @@ def compile(expr, params=None):
     str
 
     """
-    from ibis.spark.compiler import to_sql
+    from ibis.backends.spark import to_sql
 
     return to_sql(expr, dialect.make_context(params=params))
