+++ ibis/backends/spark/client.py
@@ -11,10 +11,10 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.client import Database, Query, SQLClient
-from ibis.spark import compiler as comp
-from ibis.spark import ddl
-from ibis.spark.compiler import SparkDialect, build_ast
-from ibis.spark.datatypes import spark_dtype
+from ibis.backends.spark import compiler as comp
+from ibis.backends.spark import ddl
+from ibis.backends.spark.compiler import SparkDialect, build_ast
+from ibis.backends.spark.datatypes import spark_dtype
 from ibis.util import log
