+++ ibis/backends/spark/ddl.py
@@ -5,7 +5,7 @@ from ibis.impala.ddl import (  # noqa: F401
     _is_quoted,
     fully_qualified_re,
 )
-from ibis.spark.compiler import _type_to_sql_string
+from ibis.backends.spark.compiler import _type_to_sql_string
 
 _format_aliases = {'TEXTFILE': 'TEXT'}
