+++ ibis/backends/spark/tests/test_client.py
