+++ ibis/backends/spark/tests/test_ddl.py
