+++ ibis/backends/spark/tests/test_numeric.py
