+++ ibis/backends/spark/tests/test_udf.py
@@ -13,7 +13,7 @@ pytestmark = [pytest.mark.spark, pytest.mark.udf]
 
 py4j = pytest.importorskip('py4j')
 ps = pytest.importorskip('pyspark')
-from ibis.spark.udf import udf  # noqa: E402, isort:skip
+from ibis.backends.spark.udf import udf  # noqa: E402, isort:skip
 
 
 @pytest.fixture(scope='session')
@@ -147,7 +147,7 @@ my_string_length_fns = [my_string_length, my_string_length_pandas]
 
 
 def test_spark_dtype_to_ibis_dtype():
-    from ibis.spark.datatypes import _SPARK_DTYPE_TO_IBIS_DTYPE
+    from ibis.backends.spark import _SPARK_DTYPE_TO_IBIS_DTYPE
 
     assert len(_SPARK_DTYPE_TO_IBIS_DTYPE.keys()) == len(
         set(_SPARK_DTYPE_TO_IBIS_DTYPE.values())
