+++ ibis/backends/spark/udf.py
@@ -14,8 +14,8 @@ import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.signature as sig
 import ibis.udf.validate as v
-from ibis.spark.compiler import SparkUDAFNode, SparkUDFNode, compiles
-from ibis.spark.datatypes import spark_dtype
+from ibis.backends.spark.compiler import SparkUDAFNode, SparkUDFNode, compiles
+from ibis.backends.spark.datatypes import spark_dtype
 
 _udf_name_cache = collections.defaultdict(itertools.count)
